Find element by id

+++ src/main/java/com/caspercodes/pokemonapi/repository/PokemonRepository.java
@@ -3,5 +3,5 @@ package com.caspercodes.pokemonapi.repository;
 import com.caspercodes.pokemonapi.model.Pokemon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PokemonRepository extends JpaRepository<Pokemon, Integer> {
+public interface PokemonRepository extends JpaRepository<Pokemon, Long> {
 }
